Hunter: update to v0.24.9 for OpenBLAS v0.3.21
Hunter v0.24.9 got the update of OpenBLAS v0.3.21. This includes a
f2c-converted copy of LAPACK if not fortran compiler is available. This
means we can compile OpenBLAS to provide BLAS and LAPACK functionality
with a regular C++ compiler like MSVC or GCC. This includes compiling
everything as a static library. If that library is later used to compile
an executable the executable can run without any external dependencies
(like fortran runtimes or lapack dlls)