Add option WITH_OPENBLAS to replace generic BLAS and LAPACK
OpenBLAS `v0.3.21` now ships with an f2c-converted copy of LAPACK `v3.9.0`
which is used if no fortran compiler is available. This
means we can compile OpenBLAS to provide BLAS and LAPACK functionality
with a regular C++ compiler like MSVC or GCC. This includes compiling
everything as a static library. If that library is later used to compile
an executable the executable can run without any external dependencies
(like fortran runtimes or lapack dlls in the generic fortran based BLAS
and LAPACK implementation)

When building OpenBLAS yourself be sure to use 'BUILD_WITHOUT_LAPACK=NO' and
mabye with 'NOFORTAN=1' to get a pure C++ OpenBLAS library (with the
benefits described above)

Add section in README.md to reflect the mentioned changes and new
`WITH_OPENBLAS` option.