CI: add github-actions based jobs
Use the GitHub provided GitHub-Actions to run MSVC and Ubuntu CI jobs.

The CI jobs trigger on every push (on each owns repository), and on pull
requests against the `master` branch.

In the process update to test C++11 and C++20 standards (instead of
C++17).

The clang job builds without Metis support as it errors with an `-fPIC`
related error message.

The linux-mingw based jobs are there, but disabled as they break with
'multiple definition' error.