Add option `METIS_IDXTYPEWIDTH` with default 64
When building ceres-solver 2.2.0 with SuiteSparse+Metis and EigenSparse
support the build errors because the definition of
`#define IDXTYPEWIDTH 64` in `metis.h` and the hard coded index type of
`ceres-solver` of `int`.

Add a CMake configuration option `METIS_IDXTYPEWIDTH` to override the
`IDXTYPEWIDTH` with 32 again to be compatible with ceres-solver 2.2.0.
The default value is kept to be `64` to not change the default behavior,
but add the possibility to override the change back to `32`.

Fixes: https://github.com/jlblancoc/suitesparse-metis-for-windows/issues/116