suitesparse-metis: Update project as metis now part of CHOLMOD
The upstream SuiteSparse 7.5.1 project now includes the metis project as
a subfolder of the `CHOLMOD` module with the name `SuiteSparse_metis`.
This copy is modified by the project to the point, that external metis
versions probably won't work anymore and the included copy only works
for the `CHOLMOD` module.

Therefore use Metis as private dependency and remove the
`SuiteSparse::metis` target from the installed targets.

The following changes have been done to the cmake project files:
- SuiteSparse: Metis is CHOLMOD private dependency now, don't install
- SuiteSparse: metis include only needed for CHOLMOD
- cholmod: glob c files, remove wrappers, use metis directly
- SPQR: remove metis linkage
- config: remove export of SuiteSparse::metis target, internal only
- remove SuiteSparse_LINKER_METIS_LIBS and SuiteSparse_EXPORTED_METIS_LIBS