CXSparse: disable complex usage in cs.h
Since the upgrade to SuiteSparse 7.5.1 the `cs.h` file in upstream is
created by CMake from the `cs.h.in` template. To limit the CMake code
just use the checked in `cs.h`.

But the checked in `cs.h` has complex support enabled by default, which
is a problem for `MSVC` builds. So for now disable the complex support
in the checked in `cs.h` and also remove the possibility to change it
from CMake by options.

The following changes have been done to the CMake project:
- CXSparse: disable complex usage in cs.h
- cmake: CXSparse: use c files directly, not SourceWrappers
- cmake: CXSparse: no complex support, hard coded
- cmake: SuiteSparse: no complex support, remove option